refactor(synthetic): move generation out of eval
