+++ src/sampleworks/eval/generate_synthetic_density.py
@@ -9,7 +9,7 @@ from typing import ClassVar
 import torch
 from loguru import logger
 from sampleworks.core.forward_models.xray.real_space_density import XMap_torch
-from sampleworks.eval.synthetic_utils import (
+from sampleworks.synthetic.synthetic_utils import (
     load_structure_for_synthetic_reward,
     resolve_parallel_jobs,
     validate_occupancy_values,
