+++ src/sampleworks/synthetic/__init__.py
@@ -1,0 +1,1 @@
+"""Synthetic data-generation utilities."""
